Add awful.util.spawn_with_lines
This spawns a program, similarly to awful.spawn, but captures its standard
output. On each line of output, a Lua function is called with this line. When
the standard output is closed, another function is called. The intention for
this second function is "the program is done", because most programs should only
close their standard output when they exit.

Signed-off-by: Uli Schlachter <psychon@znc.in>